refactor(platform): extract host bundle selection
